Updated some docs, made many constants doubles instead of ints

+++ core/src/main/java/com/first1444/frc/robot2020/SwerveSetup.java
@@ -23,7 +23,7 @@ public interface SwerveSetup {
     /** @return track width in meters*/
     double getTrackWidth();
 
-    int getQuadCountsPerRevolution();
+    double getQuadCountsPerRevolution();
 
     MutableValueMap<ModuleConfig> setupFL(MutableValueMap<ModuleConfig> config);
     MutableValueMap<ModuleConfig> setupFR(MutableValueMap<ModuleConfig> config);
